requirements.txt
streamlit>=1.30.0
tensorflow>=2.15.0
pandas
Pillow
requests
matplotlib
scikit-learn>=1.4.0